pyproject.toml
[project]
name = "origin-miner"
version = "0.1.0"
description = "A Git forensics tool to audit commit velocity and volume."
readme = "README.md"
authors = [
    { name = "Joe", email = "joe@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    "gitpython>=3.1.7",
    "rich>=13.7.0",
]

[project.scripts]
miner-audit = "miner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# FIX 1: Tell Hatch to include miner.py directly
[tool.hatch.build.targets.wheel]
packages = ["miner.py"]

# FIX 2: Use the non-deprecated dependency group syntax
[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
